+++ include/common/mir/executor.h
@@ -57,6 +57,47 @@ class NonBlockingExecutor: public Executor
 {
 };
 
+class ThreadPoolExecutor : public NonBlockingExecutor
+{
+public:
+    void spawn(std::function<void()>&& work) override;
+
+    /**
+     * Set a handler to be called should an unhandled exception occur on the ThreadPoolExecutor
+     *
+     * By default the ThreadPoolExecutor will allow exceptions to propagate, resulting in process termination.
+     *
+     * \param handler [in]  A handler that will be called in exception context should an exception escape
+     *                      any functor executed by \ref spawn.
+     *                      This handler will be called on an unspecified thread.
+     */
+    static void set_unhandled_exception_handler(void (*handler)());
+
+    /**
+     * Wait for all current work to finish and terminate all worker threads
+     */
+    static void quiesce();
+protected:
+    ThreadPoolExecutor() = default;
+};
+
+extern NonBlockingExecutor& thread_pool_executor;
+
+/**
+ * An Executor that makes the following concurrency guarantees:
+ *
+ * 1.   If linearising_executor.spawn(A) happens-before linearising_executor.spawn(B) then
+ *      A() happens-before B(). That is, the entire execution of A is completed before B is started.
+ * 2.   No work is performed concurrently. For any two calls linearising_executor.spawn(A)
+ *      linearising_executor.spawn(B) either A() happens-before B() or B() happens-before A().
+ * 3.   Work is deferred; linearising_executor.spawn(A) will not block on the execution of A
+ */
+extern NonBlockingExecutor& linearising_executor;
+
+/**
+ * An Executor that runs work on the current thread within spawn()
+ */
+extern Executor& immediate_executor;
 }
 
 #endif // MIR_EXECUTOR_H_
