+++ src/common/thread_pool_executor.cpp
@@ -14,7 +14,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "mir/thread_pool_executor.h"
+#include "mir/executor.h"
 
 #include "mir/thread_name.h"
